Make " in a track title work in Leaflet and OpenLayers.
I need to dispaly an icon next to a track in layerswitcher, like for
markers. Therefore I add in the track title something like "<img
src="/fileadmin/bla.png">Track1"
The current implementation breaks these tracks either in OpenLayers or
in Leaflet. I changed the way how Leaflet includes tracks so that the
same track works in OpenLayers and Leaflet without causing JavaScript
errors.

This is a breaking change if someone was using such a hack in the tracks
title field with Leaflet before.
Maybe we could add an icon field for tracks to be displayed in
layerswitcher like we have for markers.