Include geoPHP as dependancy via composer.
If you add a track record (.kml, .gpx, ...) the 3rd party library geoPHP
is used. This has been included by this extension in the folter
Resources/public/geoPHP.

The current inclusion in TceMain.php did not work with TYPO3 9.5
anymore. See #42. Nevertheless this is not the clean way nowadays.

This patch adds phayes/geophp as requirement in composer.json. This way,
the geoPHP class is available in TceMain.php as well.