Fix namespace of external class geoPHP.
If you add a track record (.kml, .gpx, ...) the 3rd party library geoPHP
is used. This has been included by this extension in the folter
Resources/public/geoPHP.

Including this way needs to specify a namespace now. This is fixed in
this patch.

Additionally the used external library is added as requirement in
composer.json. In case you install ods_osm via composer, also
phayes/geophp is installed and used via composer. If you install the
extension the classic way, the local copy of geoPHP is used instead.