[BUGFIX] Add flexform manipulation to ensure valid allowed tables.
The flexform field "marker" use the group element (`type="group"`) to
reference to marker, addresses, tracks, etc.

The option `allowed` keeps a list of allowed tables for the element
browser. This browser makes it easy to select the records.

Unfortunately, if a table does not exist, there is thrown an exception
in TYPO3 11.

This flexform manipulation makes sure to always have to proper list of
allowed tables. Of course, this assumes, that all installed extensions
have created the corresponding tables.